[refactor] Segement package in project
- com.naver.commerce_platform.junit 패키지에 세부 패키지 추가 및 파일
이동
 - .test <- Calculator.java 이동
 - .utils.annotation <- Repeat.java 이동
 - .utils.info <= RepeatTestInfo.java 이동
 - .utils.listener <- RepeatRunListener.java 이동
 - .utils.runner <- RepeatRunner.java 이동
