+++ src/test/java/com/naver/commerce_platform/junit/utils/annotation/Repeat.java
@@ -1,4 +1,4 @@
-package com.naver.commerce_platform.junit;
+package com.naver.commerce_platform.junit.utils.annotation;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
